Back/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
python-dotenv
drf-spectacular
django-ckeditor